refactor: eliminate util.rs and consolidate helper functions
- Move distance_to_space and within_bounding_box from src/util.rs to src/kdtree.rs
- Rename within_bounding_box to in_bounding_box and make it private
- Organize helper functions for top-down readability flow
- distance_to_space placed after nearest_step (its user) in NEAREST HELPERS
- in_bounding_box placed after bounding_box (its user) in BOUNDING BOX
- Remove util module declaration from src/lib.rs
- Migrate tests from util.rs to kdtree.rs test module
- Remove unnecessary doc comments from private helper functions
